Extract sidebar sub-components from App.vue
Pull the five distinct sidebar regions into standalone components
(SidebarHeader, RoleSelector, SessionTabBar, SuggestionsPanel,
ChatInput) so each can be reused / repositioned independently. This
is preparation for the planned Stack-mode layout where toolbars and
input move outside the left sidebar; doing the extraction separately
keeps that layout change to pure markup movement.

Each component owns its own local state and click-outside wiring
where applicable. App.vue drops ~460 lines and a handful of refs
that only existed to thread DOM handles to top-level click-outside
composables.

- RoleSelector: owns its dropdown open-state + click-outside
- SessionTabBar: exposes the history-button ref for popup positioning
- SuggestionsPanel: absorbs the useQueriesPanel composable (deleted)
- ChatInput: absorbs file paste/drop, IME, expanded-editor modal
- SidebarHeader: owns LockStatusPopup state + its click-outside

No UI change intended. All existing data-testids preserved.
